// src/app/open-api/generated/model/document-type.ts
/**
 * GED API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DocumentColumn } from './document-column';


export interface DocumentType { 
    readonly id?: number;
    name: string;
    /**
     * Template string defining how leaf nodes should be displayed in the document hierarchy tree. Supports placeholders such as {name}, {id}, or any column name.
     */
    leafFormat?: string;
    treeViewHierarchy?: Array<string>;
    columns?: Array<DocumentColumn>;
}